Fixed error after widget cancellation
